+++ k8sutil/pkg/app/webhookcert/reconciler_test.go
@@ -50,7 +50,8 @@ func TestReconcilerUpdatesCABundle(t *testing.T) {
 			}()
 
 			mgr, err := manager.New(eit.RESTConfig, manager.Options{
-				Namespace: ns.GetName(),
+				Namespace:          ns.GetName(),
+				MetricsBindAddress: "0",
 			})
 			require.NoError(t, err)