+++ k8sutil/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+* A new application, portforward, supports programmatic port forwarding to
+  arbitrary Kubernetes pods and services.
+
 ## [0.6.7] - 2022-03-23
 
 ### Added